Get rid of mentions of state in model class.
Model class should only be in charge of the model. State management can be done by the calling code. This way the model doesn't overfit on a single type of memory. Will be able to use any memory class following the same schema. Which could allow us to swap in other more interesting memory caches.