First attempts at tuning model to beat level 1
Noticed a few issues with the model. Currently in the process of exploring + addressing them.

Noted issues:
- Gradients going to 0 (was mostly due to having 0 reward sometimes). Addressed by adding penalties (losnig game, too long without a reward)
- Memory issues. Due to long episodes and training on the "full batch"
- Overfitting? Think the model might be learning one "path" and then overfitting on it. Still need to address this one.
- Spiky rewards. Model doesn't have smooth increase in performance. It varies widly episode to episode. Adjusting learning rates to see if that helps.
- Changed the optimizer.
- NaN errors. Caused by division by 0 because of the 0 reward. Solved with the penalties.

Still in the process of trying to adjuse and improve the model so it can beat level 1. Once I do that will use DQN or some other "well known" model / approach and see what the diffs are.